Make all_training_net a model property, warmup train_step twice
Once, to get training_nets per routine/model (nets are run in eval mode)
Second, to warmup buffers for each training net (e.g. for BatchNorm)

Warmed up buffers are needed if an eval_step is first done due to
'validation stage' of a model's train_loop. This is desired for the
'autotune stage' or optimizers' schedulers.